Move the moderate card to the shared components for admin
CORL-443
