+++ src/core/client/admin/components/ModerateCard/CommentContent.css
