+++ src/core/client/admin/components/ModerateCard/CommentContent.spec.tsx
