+++ src/core/client/admin/components/ModerateCard/CommentContent.tsx
