+++ src/core/client/admin/components/ModerateCard/FlagDetailsCategory.css
@@ -1,0 +1,3 @@
+.category {
+  color: var(--palette-error-darkest);
+}
