+++ src/core/client/admin/components/ModerateCard/FlagDetailsCategory.tsx
