+++ src/core/client/admin/components/ModerateCard/FlagDetailsContainer.tsx
