+++ src/core/client/admin/components/ModerateCard/FlagDetailsEntry.tsx
