+++ src/core/client/admin/components/ModerateCard/ModerateCard.spec.tsx
