+++ src/core/client/admin/components/ModerateCard/index.ts
